Prepare release v0.8.0

+++ src/bartz/mcmcstep/_step.py
@@ -1342,7 +1342,7 @@ def apply_moves_to_split_trees(
 
 @jax.jit
 def _sample_wishart_bartlett(
-    key: Key[Array, ''], df: Integer[Array, ''], scale_inv: Float32[Array, 'k k']
+    key: Key[Array, ''], df: Float32[Array, ''], scale_inv: Float32[Array, 'k k']
 ) -> Float32[Array, 'k k']:
     """
     Sample a precision matrix W ~ Wishart(df, scale_inv^-1) using Bartlett decomposition.
